x86_64 compile error fixes
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@400 048eaad1-c872-4173-8c30-523bb29f53ec

+++ crc.c
@@ -95,7 +95,7 @@ crc
 crcSlow(unsigned char const message[], int nBytes)
 {
     crc            remainder = INITIAL_REMAINDER;
-	signed long int byte;
+	int byte;
 	unsigned char  bit;
